Add support for usets as incremental node values
The shape node becomes a computation-node, with the usets used by its
effect as named inputs. Add a "predraw queue" to the render.
(invalidate-calculation shape t t) pushes a function on this queue to
update the environment, possibly forcing an upload of the uset.

Change simple-environment to take uset names, not usets, as arguments.

Rework cube.lisp to use computation nodes.