Optimize VAO creation
Add support for interleaved attributes; it isn't practical to allocate
multiple shape attributes from a single buffer if they aren't
interleaved. Add BUFFER-OFFSET to MIRRORED-RESOURCE, which is now the
actual location of the mirrored data in buffer memory; OFFSET is changed
to mean the pointer value used in the call to gl:vertex-attrib-pointer.