Throttle motion events based on time
Previously motion events were dropped until they stopped coming i.e.,
the window system's queue was drained, but this didn't work on some systems;
on Arch ARM Linux, for example, the queue is never drained until the
user stops moving the mouse. Also, glop doesn't provide access to the
motion hint option in X, and I don't know if that's even portable. So,
now we deliver motion events based on the time of last update and a
target framerate of the application. This whole subject of a flexible
real-time event loop needs more thought.