Reimplement incremental nodes with MOP
Keywords :input-reader, :input-writer, :input-accessor, and
:compute-function are now supported in the slot definitions of classes
with metaclass compute-class. A connect function is used to hook nodes
together.

Also, input slots are now part of an effect, not a bundle. As a result,
a new effect needs to be instantiated for each bundle. The examples now
use factory classes to manage the creation of effects and the sharing of
their graphics state.